Added persistence to helm chart

+++ src/mocktrics_exporter/metricCollection.py
@@ -29,7 +29,7 @@ class MetricsCollection:
             metaMetrics.metrics.metric_created.inc()
         self.update_metrics()
         logging.info(f"Adding metric: {id}: {metric}")
-        if not read_only:
+        if not read_only and persistence.database is not None:
             persistence.database.add_metric(metric)
 
         return id
@@ -37,10 +37,10 @@ class MetricsCollection:
     def add_metric_value(self, id: str, value: MetricValue) -> None:
         metric = [metric for metric in self._metrics if metric.name == id][0].metric
         metric.add_value(value)
-
-        persistence.database.add_metric_value(
-            value, persistence.database.get_metric_id(metric.name)
-        )
+        if persistence.database is not None:
+            persistence.database.add_metric_value(
+                value, persistence.database.get_metric_id(metric.name)
+            )
 
     def get_metrics(self) -> list[Metric]:
         return [metric.metric for metric in self._metrics]
@@ -58,14 +58,16 @@ class MetricsCollection:
         metaMetrics.metrics.metric_deleted.inc()
         self.update_metrics()
         logging.info(f"Removing metric: {id}: {metric.name}")
-        persistence.database.delete_metric(metric.metric)
+        if persistence.database is not None:
+            persistence.database.delete_metric(metric.metric)
 
     def delete_metric_value(self, id: str, labels: list[str]) -> None:
         metric = [metric for metric in self._metrics if metric.name == id][0].metric
         for value in metric.values:
             if all([label in value.labels for label in labels]):
                 metric.values.remove(value)
-                persistence.database.delete_metric_value(metric, value)
+                if persistence.database is not None:
+                    persistence.database.delete_metric_value(metric, value)
                 break
 
     def update_metrics(self) -> None:
